convert a couple notifs to use redux
